+++ Reference/README.md
@@ -1,7 +1,7 @@
 referenced resource:
 
 [llvm doc](https://llvm-tutorial-cn.readthedocs.io/en/latest/)
-
+[llvm doc](https://kaleidoscope-llvm-tutorial-zh-cn.readthedocs.io/zh_CN/latest/chapter-3.html#id9)
 [lex&bison win-version]()
 
 some useful solution:
