+++ Src/src/util.cpp
@@ -1,0 +1,28 @@
+#include "../include/util.h"
+map<int,string> typeMap;
+map<int,string> opMap;
+void setMap()
+{
+    typeMap[VALUEBOOL] = "Bool";
+    typeMap[VALUEINT] = "Int";
+    typeMap[VALUEFLOAT] = "Float";
+    typeMap[VALUEVOID] = "Void";
+    opMap[OPADD] = "+";
+    opMap[OPSUB] = "-";
+    opMap[OPDIV] = "/";
+    opMap[OPMUL] = "*";
+    opMap[OPMOD] = "%";
+    opMap[OPPOW] = "^";
+    opMap[OPEQ] = "==";
+    opMap[OPGT] = ">";
+    opMap[OPLT] = "<";
+}
+
+string& getTypeMap(int index)
+{
+    return typeMap[index];
+}
+
+string& getOpMap(int index){
+    return opMap[index];
+}
